requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
dj-database-url==2.1.0
Django==5.2.3
django-cors-headers==4.7.0
django-ratelimit==4.1.0
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==21.2.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.9
Pygments==2.19.2
PyJWT==2.9.0
python-decouple==3.8
redis==6.2.0
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.6.0
